Move the generateOauthClientToken function into the utils file
