+++ server/utils/utils.ts
@@ -5,6 +5,7 @@ import { Person } from '../@types/Person'
 import { RoshData } from '../@types/Rosh'
 import { NewStep, StepStatus } from '../@types/StepType'
 import { GoalStatus } from '../@types/GoalType'
+import config from '../config'
 
 const properCase = (word: string): string =>
   word.length >= 1 ? word[0].toUpperCase() + word.toLowerCase().slice(1) : word
@@ -166,4 +167,12 @@ export function sortSteps(steps: NewStep[]) {
 
 export function goalStatusToTabName(status: GoalStatus): string {
   return status === GoalStatus.ACTIVE ? 'current' : status.toLowerCase()
+}
+
+export default function generateOauthClientToken(
+  clientId: string = config.apis.hmppsAuth.apiClientId,
+  clientSecret: string = config.apis.hmppsAuth.apiClientSecret,
+): string {
+  const token = Buffer.from(`${clientId}:${clientSecret}`).toString('base64')
+  return `Basic ${token}`
 }
